Make dependency optional
The glib2 tool is only required when schema should be processed. To
shorten the configuration and avoid confusing the waf script user with
unneeded options, it should be able to be left out.

Passing the responsibility to load the tool to the waf script author
cannot be avoided: Whether glib2 would need to be automatically leaded
can only be decided when a schema is encountered in the build stage
while loading has to be already decided while configuring.