Change inclusion pattern
JavaScript includes can come in a variety of shapes. Allowing for
dynamic change of the pattern enables use of the inclusion scanner with
arbitrary formats.

Instead of mandating a group containing the input, the whole match could
arguably be used and still find the sought pattern using look around
expressions. As some of these have to be of fixed length and the pattern
surrounding the group do not, the latter approach proves more flexible.