Postpone feature processing to task
By itself, this does nothing more than complicate the task. But as a
task, the process can have dependencies, most importantly the source and
meta data files that are scanned. They may be generated during the build
and can thus be missing while the generator posts, but available to a
task dependent upon them. This commit establishes the necessary
structure of a task creating new tasks by invoking a generator.

As it does not make any predictions about the inputs and outputs of the
tasks it creates and even emits installation directives, the task must
always run. Further dependency resolution is left to the created tasks.