Eradicate internal feature
Features are exposed by nature. Thus, an internal feature is a
contradiction in itself. Including all of its content into the task
which before solemnly executes it really internalizes the code to begin
with. It also allows for some variable recycling.

A generator still has to be instantiated as this is the way the glib2
tool processes schemas.

This commit is best viewed ignoring whitespace, passing the -w switch to
most of Git's patch view displaying commands.