Only create installation tasks when installing
The add_install_files function will not have any effect to begin with
when the context only builds and does not install. So in this case, the
categorization of prerequisites necessary only to compose its parameters
can be skipped entirely.

This commit moves around many code parts to collect all skipped content
underneath one new if predicate, which is the only really new
instruction herein.