Distinguish scanned sources
The scanning process can now do more than the scanning function
available to the wscript author: It can scan files that are first
created during the build. To employ this feature for manually specified
files, the sources are now also scanned.

A new parameter for data files, is provided retaining the previous
functionality of adding files to install that need not be scanned.